refactor : playlist 패키지 정리
