+++ src/main/java/com/mopl/api/domain/playlist/repository/PlaylistRepository.java
@@ -1,8 +1,7 @@
 package com.mopl.api.domain.playlist.repository;
 
 import com.mopl.api.domain.playlist.entity.Playlist;
-import com.mopl.api.domain.playlist.entity.PlaylistContent;
-import java.util.Optional;
+import com.mopl.api.domain.playlist.repository.impl.PlaylistRepositoryCustom;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
